+++ back-end/src/main/java/com/devminds/rentify/web/UserController.java
@@ -17,7 +17,6 @@ public class UserController {
 
     @PostMapping("/google-login")
     public void handleGoogleLogin(@RequestHeader("Authorization") String googleCredential) {
-
     }
